+++ app/routers/service_routers.py
@@ -5,7 +5,7 @@ from entities.service_name.model_service_name import ServiceNameModel
 from routers.consts import RouteSlug
 from models.database import get_session
 from services.service_service import ServiceNameService
-from schemas.service_name_schema import ServiceNameOptionalSchema
+from entities.service_name.schema_service_name import ServiceNameOptionalSchema
 from dependencies.base_dependency import (
     ValidGetByIdDependency,
     valid_empty_schema,
